improved reliablity of mange_dev python script and added bash wrapper

+++ streamwebs_frontend/manage_dev_accounts.py
@@ -1,24 +1,30 @@
 import os
 import sys
 
+# Check the input arguments to decided action.
+# DEFAULT: create
 action = ''
 if len(sys.argv) < 2:
     action = 'create'
 else:
     action = sys.argv[1]
 
+# If action was 'help' print out script syntax and exit
 if action == 'help' or action == '--help':
     print('Usage: %s [create|school|delete] [school_name]' % sys.argv[0])
     sys.exit(0)
 
+# Load the django instance
 from django.core.wsgi import get_wsgi_application
 from django.core.exceptions import ObjectDoesNotExist
 
+# Point django to correct folders and files
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "streamwebs_frontend.settings")
 proj_path = "."
 sys.path.append(proj_path)
 application = get_wsgi_application()
 
+# Import models for editing
 from django.contrib.auth.models import User, Group
 from streamwebs.models import School, UserProfile
 
@@ -30,6 +36,7 @@ super_admin_name = 'superadmin'
 org_admin_name = 'teacher'
 org_author_name = 'student'
 
+# Check to see if there are any schools in the database at all
 if School.objects.all().exists():
     default_school = School.objects.order_by('name').first()
 else:
@@ -37,12 +44,15 @@ else:
     sys.exit(1)
 
 # Create action
+# Description: creates all the dev accounts and profiles
 if action == 'create':
 
+    # Pull the school name from the command line or use the default school
     if len(sys.argv) < 3:
         school = default_school
     else:
-        school = School.objects.filter(name=sys.argv[2]).first()
+        school_name = ' '.join(sys.argv[2:])
+        school = School.objects.filter(name=school_name).first()
         if school is None:
             print('ERROR: couldn\'t find the school \'%s\'' % sys.argv[2])
             sys.exit(2)
@@ -139,11 +149,16 @@ if action == 'create':
             else:
                 print('\'' + org_author_name + '\' user exists, linked to school \'' + tmp_school.name + '\'')
 
+# School action
+# Description: will change the org_admin's and org_author's school link
 elif action == 'school':
+
+    # Pull the shcool name from the arguments, else use the default school
     if len(sys.argv) < 3:
         school = default_school
     else:
-        school = School.objects.filter(name=sys.argv[2]).first()
+        school_name = ' '.join(sys.argv[2:])
+        school = School.objects.filter(name=school_name).first()
         if school is None:
             print('ERROR: couldn\'t find the school \'%s\'' % sys.argv[2])
             sys.exit(2)
@@ -162,21 +177,36 @@ elif action == 'school':
     org_author_profile.save()
     print('\'' + org_author_name + '\' user linked to school \'' + school.name + '\'')
 
+# Delete action
+# Description: will delete all the dev accounts
 elif action == 'delete':
+
+    # Delete the super admin
     super_admin_account = User.objects.filter(username=super_admin_name).first()
-    super_admin_profile = UserProfile.objects.filter(user=super_admin_account).first()
-    super_admin_account.delete()
-    super_admin_profile.delete()
-    print('\'' + super_admin_name + '\' user deleted')
+    if super_admin_account:
+        super_admin_account.delete()
+        print('\'' + super_admin_name + '\' user deleted')
+    else:
+        print('\'' + super_admin_name + '\' user didn\'t exist')
 
+    # Delete the org admin
     org_admin_account = User.objects.filter(username=org_admin_name).first()
-    org_admin_profile = UserProfile.objects.filter(user=org_admin_account).first()
-    org_admin_account.delete()
-    org_admin_profile.delete()
-    print('\'' + org_admin_name + '\' user deleted')
+    if org_admin_account:
+        org_admin_profile = UserProfile.objects.filter(user=org_admin_account).first()
+        org_admin_account.delete()
+        if org_admin_profile:
+            org_admin_profile.delete()
+        print('\'' + org_admin_name + '\' user deleted')
+    else:
+        print('\'' + org_admin_name + '\' user didn\'t exist')
 
+    # Delete the org author
     org_author_account = User.objects.filter(username=org_author_name).first()
-    org_author_profile = UserProfile.objects.filter(user=org_author_account).first()
-    org_author_account.delete()
-    org_author_profile.delete()
-    print('\'' + org_author_name + '\' user deleted')
+    if org_author_account:
+        org_author_profile = UserProfile.objects.filter(user=org_author_account).first()
+        org_author_account.delete()
+        if org_author_profile:
+            org_author_profile.delete()
+        print('\'' + org_author_name + '\' user deleted')
+    else:
+        print('\'' + org_author_name + '\' user didn\'t exist')
